requirements.txt
streamlit>=1.8.0
python-dotenv>=0.19.0
google-generativeai>=0.1.0
PyPDF2>=2.0.0
pandas>=1.3.0
nltk>=3.6.0
plotly>=5.5.0
matplotlib>=3.4.0
seaborn>=0.11.0
wordcloud>=1.8.0
requests>=2.26.0
PyGithub>=1.55.0